remove modules & move folder around
