+++ components/common/Layout.tsx
@@ -1,4 +1,4 @@
-import Navbar from "./navigation/Navbar";
+import Navbar from "../navigation/NavBar";
 
 const Layout = ({ children }: any) => (
   <div className="min-h-screen text-white bg-primaryBackground">
