+++ components/content/ContentHeader.tsx
@@ -1,5 +1,4 @@
 import Image from "next/image";
-import { FaPlay } from "react-icons/fa";
 
 interface IContentHeader {
   cover: string;
